bootstrap: fix Darwin stat ABI for from-stage0 gcc — 64-bit-inode struct stat + ino_t + $INODE64 symbol redirect
The from-stage0 gcc-15 (phase47, HOST_BUILD_CC=0) failed building cold from
the hex0 seed because the bootstrap libc's struct stat didn't match the
64-bit-inode Darwin ABI:

  - ino_t was 32-bit, making cpp_dir/cpp_buffer too small, so genmatch
    truncated BUILT_IN_CBRTF128 ("no such operator BUILT_IN_CBR");
  - struct stat field order (st_ino before st_mode) didn't match Darwin;
  - chain binaries link the binutils ld, which (unlike Apple's linker)
    does not rewrite stat/fstat to the $INODE64 symbol variant, so they
    bound the legacy 32-bit-inode stat whose layout differs (st_mode@8 vs
    st_mode@4). genmatch's libcpp then read st_mode from the wrong bytes
    and saw a regular file (cfn-operators.pd) as a block device.

Fix: make struct stat the exact 64-bit-inode Darwin layout, keep ino_t
64-bit, and declare stat/fstat/lstat/fstatat with __asm("_NAME\$INODE64")
(gcc only; tcc keeps the bootstrap libc's syscall path). boot_stat in the
libc mirror is updated to match.

phase47 now builds cold from the 4 KB seed and its strict, no-host-clang
gcc-15 produces a bit-identical GNU Hello to phase46
(phase46_phase47_equal=yes).

+++ bootstrap/headers/tcc-darwin-bootstrap/sys/types.h
@@ -12,7 +12,7 @@ typedef int pid_t;
 typedef unsigned int uid_t;
 typedef unsigned int gid_t;
 typedef long long off_t;
-typedef unsigned int ino_t;
+typedef unsigned long long ino_t;
 typedef int dev_t;
 typedef unsigned short nlink_t;
 typedef long blkcnt_t;
